+++ ChangeLog.md
@@ -1,5 +1,13 @@
 # Changelog for polysemy-methodology
 
+## v0.1.2.0
+
+* Add `fmapMethodology`, `fmapMethodology2`, `bindMethodology` and `traverseMethodology`.
+
+## v0.1.1.0
+
+* Fix erroneous type signature in `endMethodologyTerminal`.
+
 ## v0.1.0.0
 
 * Set of domain modelling tools for polysemy.